+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.2.1] - 2024-02-16
+
+### Changed
+ - matchit::generator::merge: changed from reducer to adapter implementation
+
+### Fixed
+ - Cargo.toml contained wrong versions
+
 ## [0.2.0] - 2024-02-13
 
 ### Fixed
@@ -58,6 +66,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - ripdlt: a test tool to process DLTv1 files with rapiddlt and matchit.
 
-[unreleased]: https://github.com/pgraubner/rapiddlt/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/pgraubner/rapiddlt/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/pgraubner/rapiddlt/releases/tag/v0.2.1
 [0.2.0]: https://github.com/pgraubner/rapiddlt/releases/tag/v0.2.0
 [0.1.0]: https://github.com/pgraubner/rapiddlt/releases/tag/v0.1.0